test(providers): derive raw client guard from exports
